* commit '79c2fa33697c7fd2d76b1182ccaf3b4106a9a44b': (32 commits)
  proxy/proxy_http.c: remove compiler warning.
  slirp-android/: remove compiler warnings.
  posix-aio-compat.c: remove compiler warnings.
  disas/i386.c: remove compiler warnings.
  util/: remove compiler warnings.
  ui/: remove compiler warnings.
  net/net-android.c: remove compiler warnings.
  android/adb-qemud.c: remove compiler warning.
  android/camera: remove compiler warnings.
  qemu-log.c: Remove compiler warning.
  target-arm/translate.c: Remove compiler warnings.
  disas/mips.c: Remove compiler warning.
  target-i384/translate.c: Remove compiler warning.
  main-loop.c: Remove compiler warning.
  android/main.c: Remove compiler warning.
  hw/i386/pc.c: Remove compiler warning.
  vl-android.c: Remove compiler warning.
  android/utils/timezone.c: Remove compiler warnings.
  android/snapshot.c: Remove compiler warnings.
  android/hw-qemud.c: Remove compiler warnings.
  ...

+++ android/hw-pipe-net.c
@@ -251,7 +251,7 @@ netPipe_sendBuffers( void* opaque, const GoldfishPipeBuffer* buffers, int numBuf
     NetPipe*  pipe = opaque;
     int       count = 0;
     int       ret   = 0;
-    int       buffStart = 0;
+    size_t    buffStart = 0;
     const GoldfishPipeBuffer* buff = buffers;
     const GoldfishPipeBuffer* buffEnd = buff + numBuffers;
 
@@ -309,7 +309,7 @@ netPipe_recvBuffers( void* opaque, GoldfishPipeBuffer*  buffers, int  numBuffers
     NetPipe*  pipe = opaque;
     int       count = 0;
     int       ret   = 0;
-    int       buffStart = 0;
+    size_t    buffStart = 0;
     GoldfishPipeBuffer* buff = buffers;
     GoldfishPipeBuffer* buffEnd = buff + numBuffers;
